Fix `-Werror,-Wtautological-constant-out-of-range-compare` build error
Clang-22 fails to build with the following error:
```
stapbpf.cxx:2033:21: error: result of comparison of constant 2305843009213693951 with expression of
      type 'unsigned int' is always false [-Werror,-Wtautological-constant-out-of-range-compare]
 2033 |   if (n_active_cpus > (size_t)-1 / sizeof(struct pollfd))
      |       ~~~~~~~~~~~~~ ^ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
```
as `n_active_cpus` is of type `unsigned` (32-bit), so it should be
compared against another `unsigned`.

+++ stapbpf/stapbpf.cxx
@@ -2030,7 +2030,7 @@ perf_event_loop(pthread_t main_thread)
     = map_attrs[bpf::globals::perf_event_map_idx].max_entries;
   unsigned n_active_cpus
     = count_active_cpus();
-  if (n_active_cpus > (size_t)-1 / sizeof(struct pollfd))
+  if (n_active_cpus > (unsigned)-1 / sizeof(struct pollfd))
     fatal("Too many active CPUs for pollfd allocation\n");
   struct pollfd *pmu_fds
     = (struct pollfd *)malloc(n_active_cpus * sizeof(struct pollfd));
